third_party/blink/web_tests/http/tests/devtools/elements/styles-4/styles-update-links-4-expected.txt

Tests that links are updated property when editing pseudo element property.


Running: selectNodeWithPseudoElements

Running: testPseudoSectionPropertyEdit


#### AFTER PROPERTY INSERTED ####


[expanded] 
element.style { ()

[expanded] 
section { (user agent stylesheet)
    display: block;

======== Inherited from html ========
[expanded] 
html { (user agent stylesheet)
    color: -internal-root-color;

======== Pseudo ::before element ========
[expanded] 
#pseudo::before { (<style>)
    color: blue;

======== Pseudo ::after element ========
[expanded] 
#pseudo::after { (<style>)
    border: 1px solid black;

[expanded] 
#pseudo::after { (<style>)
/-- overloaded --/     pseudo-property: "12";
/-- overloaded --/     PROPERTY: INSERTED;
    color: red;

0 rule ranges are equal.
0 rule ranges are equal.
3 rule ranges are equal.
3 rule ranges are equal.
5 rule ranges are equal.
